Add instructions to change Jupyter Notebook home dir
When opening Jupyter Notebook, it will by default open C:\ in the
Dashboard tree. Users won't have any files in C:\, and to be able
to navigate to their files they'll need to tell Notebook to show
N:\ as their home dir. This is as easy as first cd to N: in
Anaconda Prompt before opening jupyter notebook, but still need
instructions to explain this. Also includes instruction to make
this persist by changing a notebook config setting.