Added fix for white and borderless app windows
The base image of LASER VMs has a default theme that presents app windows with white title bars and borders. When windows overlap it's difficult to distinguish them. This guide points users to customise their VM's theme to add colour to the title bars and borders of windows, which makes the boundaries of all windows - even when not selected - much clearer.